Dropdown: added publicAPI
